requirements.txt
tqdm
requests
bs4
pyyaml
regex